LRCI-3969 Create a base pull request job entity to share logic
