+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/job/BasePortalPullRequestJobEntity.java
@@ -15,16 +15,11 @@ import org.json.JSONObject;
  * @author Michael Hashimoto
  */
 public abstract class BasePortalPullRequestJobEntity
-	extends BaseJobEntity implements PortalPullRequestJobEntity {
+	extends BasePullRequestJobEntity implements PortalPullRequestJobEntity {
 
 	@Override
 	public String getForwardReceiverUserName() {
 		return getParameterValue("forwardReceiverUserName");
-	}
-
-	@Override
-	public String getOriginName() {
-		return _originName;
 	}
 
 	@Override
@@ -39,42 +34,13 @@ public abstract class BasePortalPullRequestJobEntity
 	}
 
 	@Override
-	public String getSenderBranchName() {
-		return _senderBranchName;
-	}
-
-	@Override
-	public String getSenderBranchSHA() {
-		return _senderBranchSHA;
-	}
-
-	public String getSenderUserName() {
-		return _senderUserName;
-	}
-
-	@Override
 	public String getTestSuiteName() {
 		return getParameterValue("testSuiteName");
 	}
 
 	@Override
-	public String getUpstreamBranchName() {
-		return _upstreamBranchName;
-	}
-
-	@Override
-	public String getUpstreamBranchSHA() {
-		return _upstreamBranchSHA;
-	}
-
-	@Override
 	public void setForwardReceiverUserName(String forwardReceiverUserName) {
 		setParameterValue("forwardReceiverUserName", forwardReceiverUserName);
-	}
-
-	@Override
-	public void setOriginName(String originName) {
-		_originName = originName;
 	}
 
 	@Override
@@ -84,40 +50,12 @@ public abstract class BasePortalPullRequestJobEntity
 	}
 
 	@Override
-	public void setSenderBranchName(String senderBranchName) {
-		_senderBranchName = senderBranchName;
-	}
-
-	public void setSenderBranchSHA(String senderBranchSHA) {
-		_senderBranchSHA = senderBranchSHA;
-	}
-
-	public void setSenderUserName(String senderUserName) {
-		_senderUserName = senderUserName;
-	}
-
-	@Override
 	public void setTestSuiteName(String testSuiteName) {
 		setParameterValue("testSuiteName", testSuiteName);
-	}
-
-	public void setUpstreamBranchName(String upstreamBranchName) {
-		_upstreamBranchName = upstreamBranchName;
-	}
-
-	public void setUpstreamBranchSHA(String upstreamBranchSHA) {
-		_upstreamBranchSHA = upstreamBranchSHA;
 	}
 
 	protected BasePortalPullRequestJobEntity(JSONObject jsonObject) {
 		super(jsonObject);
 	}
-
-	private String _originName;
-	private String _senderBranchName;
-	private String _senderBranchSHA;
-	private String _senderUserName;
-	private String _upstreamBranchName;
-	private String _upstreamBranchSHA;
 
 }
