+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/job/PortalPullRequestJobEntity.java
@@ -10,42 +10,18 @@ import java.net.URL;
 /**
  * @author Michael Hashimoto
  */
-public interface PortalPullRequestJobEntity extends JobEntity {
+public interface PortalPullRequestJobEntity extends PullRequestJobEntity {
 
 	public String getForwardReceiverUserName();
 
-	public String getOriginName();
-
 	public URL getPortalPullRequestURL();
-
-	public String getSenderBranchName();
-
-	public String getSenderBranchSHA();
-
-	public String getSenderUserName();
 
 	public String getTestSuiteName();
 
-	public String getUpstreamBranchName();
-
-	public String getUpstreamBranchSHA();
-
 	public void setForwardReceiverUserName(String forwardReceiverUserName);
-
-	public void setOriginName(String originName);
 
 	public void setPortalPullRequestURL(URL portalPullRequestURL);
 
-	public void setSenderBranchName(String senderBranchName);
-
-	public void setSenderBranchSHA(String senderBranchSHA);
-
-	public void setSenderUserName(String senderUserName);
-
 	public void setTestSuiteName(String testSuiteName);
-
-	public void setUpstreamBranchName(String upstreamBranchName);
-
-	public void setUpstreamBranchSHA(String upstreamBranchSHA);
 
 }
