+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/job/PullRequestJobEntity.java
@@ -1,0 +1,37 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.jethr0.job;
+
+/**
+ * @author Michael Hashimoto
+ */
+public interface PullRequestJobEntity extends JobEntity {
+
+	public String getOriginName();
+
+	public String getSenderBranchName();
+
+	public String getSenderBranchSHA();
+
+	public String getSenderUserName();
+
+	public String getUpstreamBranchName();
+
+	public String getUpstreamBranchSHA();
+
+	public void setOriginName(String originName);
+
+	public void setSenderBranchName(String senderBranchName);
+
+	public void setSenderBranchSHA(String senderBranchSHA);
+
+	public void setSenderUserName(String senderUserName);
+
+	public void setUpstreamBranchName(String upstreamBranchName);
+
+	public void setUpstreamBranchSHA(String upstreamBranchSHA);
+
+}
